requirements.txt
numba
imageio
pandas
numpy
tqdm
Pillow
scipy
scikit-image
imutils
opencv-python